Continue to work out kinks, re-add requirements.txt, add boot-linux-sudo

+++ sunstorm.py
@@ -9,8 +9,13 @@ import subprocess
 import shutil
 import atexit
 import tempfile
+import glob
 
-ROOT = os.path.dirname(__file__) # -> XXX: global variable
+# Global variables
+ROOT = os.path.dirname(__file__) 
+LINUX = (sys.platform == 'linux')
+
+# Append PATH
 sys.path.append(ROOT + '/src')
 os.environ['PATH'] = ((ROOT + '/bin') + ':' + os.environ.get('PATH'))
 
@@ -26,7 +31,8 @@ program_list = [
   'ldid',
   'asr64_patcher',
   'restored_external64_patcher',
-  'hdiutil',
+  # `hfsplus` comes from libdmg-hfsplus
+  'hfsplus' if LINUX else 'hdiutil'
 ]
 
 def print_error(string):
@@ -39,30 +45,48 @@ def print_info(string):
 def cleanup(directory) -> None:
     """ Remove any temp-files created """
     try: 
+      if (0): # DEBUG: stop on-error removal
+        return
+
       shutil.rmtree(directory, ignore_errors=True)
     except:
       print_error(f'Failed to remove {directory} (this is a bug)')
 
 def check_for_command(prog) -> bool:
     """ Use the `command` shell-builtin to test for program """
-    return (
-      subprocess.run(f'command -v {prog} &>/dev/null', shell=True).returncode == 0
-    )
+    return bool(shutil.which(prog))
 
 def check_for_dependencies() -> None:
     """
-      Loop over {program_list}, check for every command.
+    Loop over {program_list}, check for every command.
 
-      Exits on error
+    Exits on error
     """
     for prog in program_list:
       if not check_for_command(prog):
         print_error(f'"{prog}" not found, please install it.')
         sys.exit(1)
 
+def execute(arguments) -> str:
+  """
+  Wrapper for `subprocess.run`,
+    - Stops process on error.
+    - Returns stdout on success
+
+  FIXME: There should be a better way to wrap & check status code on each command
+  """
+  # print(arguments) # DEBUG:
+
+  result = subprocess.run(arguments, capture_output=True)
+  if result.returncode > 0:
+    print_error(f'Command "{arguments}" returned non-zero')
+    sys.exit(1)
+
+  return result.stdout
+
 def prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband):
     # tempdir
-    work = tempfile.mkdtemp(suffix='restore')
+    work = tempfile.mkdtemp(prefix='restore-')
     # make a directory in the work directory called ramdisk
     os.mkdir(f'{work}/ramdisk')
     # register cleanup trap
@@ -77,65 +101,81 @@ def prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband):
     with open(f'{work}/BuildManifest.plist', 'rb') as f:
         manifest = Manifest(f.read())
     # get the ramdisk name
-    ramdisk_path = manifest.get_comp(board, 'RestoreRamDisk')
+    ramdisk_path = manifest.get_comp(boardconfig, 'RestoreRamDisk')
     if ramdisk_path == None:
         print_error("Error: BoardConfig was not recognized")
         sys.exit(1)
 
     # extract it using img4
     print_info('Extracting RamDisk')
-    subprocess.run(['img4', '-i', f'{work}/{ramdisk_path}', '-o', f'{work}/ramdisk.dmg'])
+    execute(['img4', '-i', f'{work}/{ramdisk_path}', '-o', f'{work}/ramdisk.dmg'])
 
-    # mount it using hdiutil
     print_info('Mounting RamDisk')
-    subprocess.run(['hdiutil', 'attach', f'{work}/ramdisk.dmg', '-mountpoint', f'{work}/ramdisk'])
+    if LINUX:
+      # extract files needed using hfsplus
+      os.makedirs(f'{work}/ramdisk/usr/sbin/')
+      os.makedirs(f'{work}/ramdisk/usr/local/bin/')
+      
+      execute(['hfsplus', f'{work}/ramdisk.dmg', 'extract', '/usr/sbin/asr', f'{work}/ramdisk/usr/sbin/asr'])
+      execute(['hfsplus', f'{work}/ramdisk.dmg', 'extract', '/usr/local/bin/restored_external', f'{work}/ramdisk/usr/local/bin/restored_external'])
+    else:
+      # mount it using hdiutil
+      print_info('Mounting RamDisk')
+      execute(['hdiutil', 'attach', f'{work}/ramdisk.dmg', '-mountpoint', f'{work}/ramdisk'])
 
     # patch asr into the ramdisk
     print_info('Patching ASR in the RamDisk')
-    subprocess.run(['asr64_patcher', f'{work}/ramdisk/usr/sbin/asr', f'{work}/patched_asr'])
+    execute(['asr64_patcher', f'{work}/ramdisk/usr/sbin/asr', f'{work}/patched_asr'])
 
-    # extract the ents and save it to {work}/asr_ents.plist like:     subprocess.run(['ldid', '-e', '{work}/ramdisk/usr/sbin/asr', '>', '{work}/asr.plist'])
+    # extract the ents and save it to {work}/asr_ents.plist like:     execute(['ldid', '-e', '{work}/ramdisk/usr/sbin/asr', '>', '{work}/asr.plist'])
     print_info('Extracting ASR Ents')
     with open(f'{work}/asr.plist', 'wb') as f:
-        subprocess.run(['ldid', '-e', f'{work}/ramdisk/usr/sbin/asr'], stdout=f)
+        f.write(execute(['ldid', '-e', f'{work}/ramdisk/usr/sbin/asr']))
+        f.flush()
     # resign it using ldid
     print_info('Resigning ASR')
-    subprocess.run(['ldid', f'-S{work}/asr.plist', f'{work}/patched_asr'])
+    execute(['ldid', f'-S{work}/asr.plist', f'{work}/patched_asr'])
     # chmod 755 the new asr
     print_info('Chmoding ASR')
-    subprocess.run(['chmod', '-R', '755', f'{work}/patched_asr'])
+    execute(['chmod', '-R', '755', f'{work}/patched_asr'])
     # copy the patched asr back to the ramdisk
     print_info('Copying Patched ASR back to the RamDisk')
-    subprocess.run(['cp', f'{work}/patched_asr', f'{work}/ramdisk/usr/sbin/asr'])
+    execute(['cp', f'{work}/patched_asr', f'{work}/ramdisk/usr/sbin/asr'])
 
     if legacy:
         print_info('Legacy mode, skipping restored_external')
     else:
         # patch restored_external 
         print_info('Patching Restored External')
-        subprocess.run(['restored_external64_patcher' ,f'{work}/ramdisk/usr/local/bin/restored_external' ,f'{work}/restored_external_patched'])
+        execute(['restored_external64_patcher', f'{work}/ramdisk/usr/local/bin/restored_external', f'{work}/restored_external_patched'])
         #resign it using ldid
         print_info('Extracting Restored External Ents')
         with open(f'{work}/restored_external.plist', 'wb') as f:
-            subprocess.run(['ldid', '-e', f'{work}/ramdisk/usr/local/bin/restored_external'], stdout=f)
+            f.write(execute(['ldid', '-e', f'{work}/ramdisk/usr/local/bin/restored_external']))
+            f.flush()
         # resign it using ldid
         print_info('Resigning Restored External')
-        subprocess.run(['ldid', f'-S{work}/restored_external.plist', f'{work}/restored_external_patched'])
+        execute(['ldid', f'-S{work}/restored_external.plist', f'{work}/restored_external_patched'])
         # chmod 755 the new restored_external
         print_info('Chmoding Restored External')
-        subprocess.run(['chmod', '-R', '755', f'{work}/restored_external_patched'])
+        execute(['chmod', '-R', '755', f'{work}/restored_external_patched'])
         # copy the patched restored_external back to the ramdisk
         print_info('Copying Patched Restored External back to the RamDisk')
-        subprocess.run(['cp', f'{work}/restored_external_patched', f'{work}/ramdisk/usr/local/bin/restored_external'])
+        execute(['cp', f'{work}/restored_external_patched', f'{work}/ramdisk/usr/local/bin/restored_external'])
 
     # detach the ramdisk
     print_info('Detaching RamDisk')
-    subprocess.run(['hdiutil', 'detach', f'{work}/ramdisk'])
+    if LINUX:
+      # sys.exit(1)
+      execute(['hfsplus', f'{work}/ramdisk.dmg', 'addall' f'{work}/ramdisk'])
+    else:
+      execute(['hdiutil', 'detach', f'{work}/ramdisk'])
+
     # create the ramdisk using pyimg4
     print_info('Creating RamDisk')
-    subprocess.run([sys.executable, '-m', 'pyimg4', 'im4p', 'create', '-i', f'{work}/ramdisk.dmg', '-o', f'{work}/ramdisk.im4p', '-f', 'rdsk'])
+    execute([sys.executable, '-m', 'pyimg4', 'im4p', 'create', '-i', f'{work}/ramdisk.dmg', '-o', f'{work}/ramdisk.im4p', '-f', 'rdsk'])
     # get kernelcache name from manifest
-    kernelcache = manifest.get_comp(board, 'RestoreKernelCache')
+    kernelcache = manifest.get_comp(boardconfig, 'RestoreKernelCache')
     # extract the kernel using pyimg4 like this: pyimg4 im4p extract -i kernelcache -o kcache.raw --extra kpp.bin 
     print_info('Extracting Kernel')
 
@@ -144,10 +184,10 @@ def prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband):
     if kpp:
         extract_kernel_args += ['--extra', f'{work}/kpp.bin']
 
-    subprocess.run(extract_kernel_args)
+    execute(extract_kernel_args)
     # patch the kernel using kernel64patcher like this: Kernel64Patcher kcache.raw krnl.patched -f -a
     print_info('Patching Kernel')
-    subprocess.run(['kernel64patcher', f'{work}/kcache.raw', f'{work}/krnl.patched', '-f', '-a'])
+    execute(['Kernel64Patcher', f'{work}/kcache.raw', f'{work}/krnl.patched', '-f', '-a'])
     # rebuild the kernel like this: pyimg4 im4p create -i krnl.patched -o krnl.im4p --extra kpp.bin -f rkrn --lzss (leave out --extra kpp.bin if you dont have kpp)
     print_info('Rebuilding Kernel')
 
@@ -156,26 +196,37 @@ def prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband):
     if kpp:
       rebuild_kernel_args += ['--extra', f'{work}/kpp.bin']
 
-    subprocess.run(rebuild_kernel_args)
+    execute(rebuild_kernel_args)
 
+    # Cleanup/trim files, hack-ish-ily:
+    file_list = [fn for fn in glob.glob(f'{work}/**/*', recursive=True) if not os.path.basename(fn).endswith('im4p')]
+    for fn in file_list:
+      if fn.startswith(work) and (1): # DEBUG: -> Don't remove user files, verify
+        try:
+          os.remove(fn)
+        except:
+          pass
+
+    print_info('Moving files')
     # Done, move files to root (dirname $0)
     shutil.move(work, ROOT)
-    work = f'{ROOT}/{os.path.basename(os.path.dirname(work))}'
+    work = os.path.realpath(f'{ROOT}/{os.path.basename(work)}')
     print_info(f'Done! Files moved to "{work}"')
 
     futurerestore_args = ['futurerestore', '-t', blob, '--use-pwndfu', '--skip-blob', '--rdsk', f'{work}/ramdisk.im4p', '--rkrn', f'{work}/krnl.im4p', '--latest-sep', '--no-baseband' if skip_baseband else '--latest-baseband', ipsw]
+    futurerestore_args_string = ' '.join(futurerestore_args)
 
     print_info('You can restore the device anytime by running the following command with the device in a pwndfu state:')
-    print(futurerestore_args)
+    print(futurerestore_args_string)
 
     # write to a file to help remember
-    with open(f'{work}/restore.command') as f:
-      print(futurerestore_args, file=f)
+    with open(f'{work}/restore.command', 'wt') as f:
+      f.write(futurerestore_args_string)
 
     # Ask user if they would like to restore the device
-    ask = input('Would you like to restore now? [Yy/Nn]')
+    ask = input('Would you like to restore now? [Yy/Nn]: ')
     if ask == 'y' or ask == 'Y':
-      subprocess.run(futurerestore_args)
+      execute(futurerestore_args)
 
     # Remove the trap so it doesn't error out
     atexit.unregister(cleanup)
@@ -184,7 +235,7 @@ def prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband):
 
 def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
     # tempdir
-    work = tempfile.mkdtemp(suffix='boot')
+    work = tempfile.mkdtemp(prefix='boot-')
     # register cleanup trap
     atexit.register(cleanup, work)
 
@@ -201,6 +252,9 @@ def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
     productbuildversion = manifest.getProductBuildVersion()
     ibss_iv, ibss_key, ibec_iv, ibec_key = api.get_keys(identifier, boardconfig, productbuildversion)
 
+    if not (ibss_iv and ibss_key and ibec_iv and ibec_key):
+      print_error('Possible incorrect identifier or boardconfig')
+
     # get ibec and ibss from manifest
     print_info('Getting IBSS and IBEC')
     ibss = manifest.get_comp(boardconfig, 'iBSS')
@@ -208,31 +262,31 @@ def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
 
     # decrypt ibss like this:  img4 -i ibss -o ibss.dmg -k ivkey
     print_info('Decrypting IBSS')
-    subprocess.run(['img4', '-i', f'{work}/' + ibss, '-o', f'{work}/ibss.dmg', '-k', ibss_iv + ibss_key])
+    execute(['img4', '-i', f'{work}/' + ibss, '-o', f'{work}/ibss.dmg', '-k', ibss_iv + ibss_key])
 
     # decrypt ibec like this:  img4 -i ibec -o ibec.dmg -k ivkey
     print_info('Decrypting IBEC')
-    subprocess.run(['img4', '-i', f'{work}/' + ibec, '-o', f'{work}/ibec.dmg', '-k', ibec_iv + ibec_key])
+    execute(['img4', '-i', f'{work}/' + ibec, '-o', f'{work}/ibec.dmg', '-k', ibec_iv + ibec_key])
 
     # patch ibss like this:  iBoot64Patcher ibss.dmg ibss.patched
     print_info('Patching IBSS')
-    subprocess.run(['iBoot64Patcher', f'{work}/ibss.dmg', f'{work}/ibss.patched'])
+    execute(['iBoot64Patcher', f'{work}/ibss.dmg', f'{work}/ibss.patched'])
 
     # patch ibec like this:  iBoot64Patcher ibec.dmg ibec.patched -b "-v"
     print_info('Patching IBEC')
-    subprocess.run(['iBoot64Patcher', f'{work}/ibec.dmg', f'{work}/ibec.patched', '-b', '-v'])
+    execute(['iBoot64Patcher', f'{work}/ibec.dmg', f'{work}/ibec.patched', '-b', '-v'])
 
     # convert blob into im4m like this: img4tool -e -s blob -m IM4M
     print_info('Converting BLOB to IM4M')
-    subprocess.run(['img4tool', '-e', '-s', blob, '-m', 'IM4M'])
+    execute(['img4tool', '-e', '-s', blob, '-m', 'IM4M'])
 
     # convert ibss into img4 like this:  img4 -i ibss.patched -o ibss.img4 -M IM4M -A -T ibss
     print_info('Converting IBSS to IMG4')
-    subprocess.run(['img4', '-i', f'{work}/ibss.patched', '-o', f'{work}/ibss.img4', '-M', 'IM4M', '-A', '-T', 'ibss'])
+    execute(['img4', '-i', f'{work}/ibss.patched', '-o', f'{work}/ibss.img4', '-M', 'IM4M', '-A', '-T', 'ibss'])
 
     # convert ibec into img4 like this:  img4 -i ibec.patched -o ibec.img4 -M IM4M -A -T ibec
     print_info('Converting IBEC to IMG4')
-    subprocess.run(['img4', '-i', f'{work}/ibec.patched', '-o', f'{work}/ibec.img4', '-M', 'IM4M', '-A', '-T', 'ibec'])
+    execute(['img4', '-i', f'{work}/ibec.patched', '-o', f'{work}/ibec.img4', '-M', 'IM4M', '-A', '-T', 'ibec'])
 
     # get the names of the devicetree and trustcache
     print_info('Getting Device Tree and TrustCache')
@@ -242,12 +296,12 @@ def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
 
     # sign them like this  img4 -i devicetree -o devicetree.img4 -M IM4M -T rdtr
     print_info('Signing Device Tree')
-    subprocess.run(['img4', '-i', f'{work}/' + devicetree, '-o', f'{work}/devicetree.img4', '-M', 'IM4M', '-T', 'rdtr'])
+    execute(['img4', '-i', f'{work}/' + devicetree, '-o', f'{work}/devicetree.img4', '-M', 'IM4M', '-T', 'rdtr'])
 
     # sign them like this   img4 -i trustcache -o trustcache.img4 -M IM4M -T rtsc
     if not legacy:
         print_info('Signing Trust Cache')
-        subprocess.run(['img4', '-i', f'{work}/' + trustcache, '-o', f'{work}/trustcache.img4', '-M', 'IM4M', '-T', 'rtsc'])
+        execute(['img4', '-i', f'{work}/' + trustcache, '-o', f'{work}/trustcache.img4', '-M', 'IM4M', '-T', 'rtsc'])
 
     # grab kernelcache from manifest
     print_info('Getting Kernel Cache')
@@ -260,11 +314,11 @@ def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
     if kpp:
         extract_kernel_args += ['--extra', f'{work}/kpp.bin']
 
-    subprocess.run(extract_kernel_args)
+    execute(extract_kernel_args)
 
     # patch it like this:   Kernel64Patcher kcache.raw krnlboot.patched -f
     print_info('Patching Kernel')
-    subprocess.run(['Kernel64Patcher', f'{work}/kcache.raw', f'{work}/krnlboot.patched', '-f'])
+    execute(['Kernel64Patcher', f'{work}/kcache.raw', f'{work}/krnlboot.patched', '-f'])
 
     # convert it like this:   pyimg4 im4p create -i krnlboot.patched -o krnlboot.im4p --extra kpp.bin -f rkrn --lzss
     print_info('Converting Kernel')
@@ -273,19 +327,29 @@ def prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy):
     if kpp:
       convert_kernel_args += ['--extra', f'{work}/kpp.bin']
     
-    subprocess.run(convert_kernel_args)
+    execute(convert_kernel_args)
 
     # sign it like this:  pyimg4 img4 create -p krnlboot.im4p -o krnlboot.img4 -m IM4M
     print_info('Signing Kernel')
-    subprocess.run([sys.executable, '-m', 'pyimg4', 'img4', 'create', '-p', f'{work}/krnlboot.im4p', '-o', f'{work}/krnlboot.img4', '-m', 'IM4M'])
+    execute([sys.executable, '-m', 'pyimg4', 'img4', 'create', '-p', f'{work}/krnlboot.im4p', '-o', f'{work}/krnlboot.img4', '-m', 'IM4M'])
 
+    # Cleanup/trim files, hack-ish-ily:
+    file_list = [fn for fn in glob.glob(f'{work}/**/*', recursive=True) if not os.path.basename(fn).endswith('im4p')]
+    for fn in file_list:
+      if fn.startswith(work) and (1): # DEBUG: -> Don't remove user files, verify
+        try:
+          os.remove(fn)
+        except:
+          pass
+
+    print_info('Moving files')
     shutil.move(work, ROOT)
-    work = f'{ROOT}/{os.path.basename(os.path.dirname(work))}'
+    work = os.path.realpath(f'{ROOT}/{os.path.basename(work)}')
     print_info(f'Done! Files moved to "{work}"')
 
     # done
     print_info('You can boot the restored device anytime by running the following command with the device in a pwndfu state:')
-    print(f'{ROOT}/scripts/' + ('boot.sh' if kpp else 'boot-A10plus.sh'))
+    print(f'{os.path.realpath(ROOT + "/scripts/")}' + ('boot.sh' if kpp else 'boot-A10plus.sh'))
 
     # Remove the trap so it doesn't error out
     atexit.unregister(cleanup)
@@ -305,8 +369,8 @@ def main():
     parser.add_argument('-i', '--ipsw', help='IPSW to restore', required=True)
     parser.add_argument('-t', '--blob', help='Blob (shsh2) to use', required=True)
     parser.add_argument('-d', '--boardconfig', help='BoardConfig to use', required=True)
-    parser.add_argument('-kpp', '--kpp', help='Use Kernel Patch Protection (KPP)', required=False, action='store_true')
-    parser.add_argument('-id', '--identifier', help='Identifier to use', required=False)
+    parser.add_argument('-kpp', '--kpp', help='Use Kernel Patch Protection (KPP) (Required on devices lower than A9)', required=False, action='store_true')
+    parser.add_argument('-id', '--identifier', help='Identifier to use (ex. iPhoneX,X)', required=False)
     parser.add_argument('--legacy', help='Use Legacy Mode (iOS 11 or lower)', required=False, action='store_true')
     parser.add_argument('--skip-baseband', help='Skip Cellular Baseband', required=False, action='store_true')
     # These options cannot be used together:
@@ -318,25 +382,36 @@ def main():
     check_for_dependencies()
 
     # Cast/modify arguments here before passing
-    ipsw  = str(args.ipsw)
-    blob = str(args.blob)
+    restore = bool(args.restore)
+    boot = bool(args.boot)
+    ipsw  = os.path.realpath(args.ipsw)
+    blob = os.path.realpath(args.blob)
     boardconfig = str(args.boardconfig).lower() # lowercase board to avoid missing errors
     kpp = bool(args.kpp)
     legacy = bool(args.legacy)
     skip_baseband = bool(args.skip_baseband)
-    identifier = str(args.identifier)
+    identifier = args.identifier
 
-    if args.restore:
-        prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband)
-    elif args.boot:
-        if not identifier:
-            print_error('You need to specify an identifier')
-            sys.exit(1)
+    if not os.path.exists(ipsw):
+      print_error(f'IPSW "{ipsw}" doesn\'t exist')
+      sys.exit(1)
 
-        prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy)
+    if not os.path.exists(blob):
+      print_error(f'Blob "{blob}" doesn\'t exist')
+      sys.exit(1)
+
+    if boot and not identifier:
+      print_error('You need to specify an identifier (--identifier)')
+      sys.exit(1)
+
+    if restore:
+      prep_restore(ipsw, blob, boardconfig, kpp, legacy, skip_baseband)
+    elif boot:
+      prep_boot(ipsw, blob, boardconfig, kpp, identifier, legacy)
     else:
-        print_error('Please specify a mode')
-        sys.exit(1)
+      print_error('No mode selected (this is a bug)')
+      print(args)
+      sys.exit(1)
 
     sys.exit(0)
